tools/mcp/bluejet_mcp/requirements.txt
fastmcp>=2.11,<3
httpx>=0.27,<1
python-dotenv>=1.0,<2

